requirements.txt
# 最後にいらないものが入っていないか確認すること！

# ===== Core =====
python>=3.8

# ===== LLM / Transformer =====
torch>=1.13.1
transformers>=4.21.0
accelerate>=0.12.0
sentencepiece
protobuf

# ===== OpenAI (data generation) =====
openai>=0.27.0

# ===== Data handling =====
numpy
pandas
scipy
pyyaml
tqdm

# ===== Evaluation =====
scikit-learn

# ===== Visualization =====
matplotlib
seaborn

# ===== Tokenization / NLP utils =====
regex

# ===== Others =====
python-dotenv